requirement.txt
asgiref==3.7.2
build==1.0.3
mysql-connector-python==8.2.0
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
cryptography==41.0.7
deprecation==2.1.0
distlib==0.3.7
Django==4.2.7
django-keycloak-auth==0.9.9
djangorestframework==3.14.0
drf-keycloak-auth==0.3.0
dry-rest-permissions==0.1.10
dulwich==0.21.6
ecdsa==0.18.0
fastjsonschema==2.19.0
filelock==3.13.1
idna==3.6
installer==0.7.0
jaraco.classes==3.3.0
keyring==24.3.0
more-itertools==10.1.0
msgpack==1.0.7
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-jose==3.3.0
python-keycloak==3.7.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
rapidfuzz==3.5.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
setuptools==69.0.2
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.3
trove-classifiers==2023.11.29
tzdata==2023.3
urllib3==1.26.18
virtualenv==20.24.7
mysqlclient==2.2.0
django-cors-headers==4.3.1